chore(prompter): refactor to allow granular consumption
Break up the files to allow granular consumption of the changelog formatter / prompter

This closes #24
